refactor(pvm2): rename rv_* / pvm2_* artefacts to unprefixed names
Phase 3c: the legacy PVM is gone, so the `rv_*` prefix that
distinguished PVM2 artefacts from PVM ones is now redundant. Rename to
the natural unprefixed names so the codebase reads as a single ISA.

Files:
* `javm-exec/src/{rv_instruction,rv_interp,rv_predecode}.rs` →
  `{instruction,interp,predecode}.rs`
* `javm-transpiler/src/linker_rv.rs` → `linker.rs`
* `javm-recompiler-x86/src/rv_codegen.rs` merged into `codegen.rs`
* `javm-bench/examples/pvm2_{smoke,dump,disasm,compile_profile,rv_opcode_stats}.rs` →
  `{smoke,dump,disasm,compile_profile,opcode_stats}.rs`
* `javm-bench/benches/pvm2_bench.rs` → `bench.rs`

Types / functions:
* `RvInterpreter` → `Interpreter`
* `RvProgram` → `Program`
* `RvInst` → `Inst`
* `RvPredecode` → `Predecode`
* `compile_rv` → `Compiler::compile`
* `link_elf_rv` → `link_elf`
* `predecode_rv` → `predecode`

Pure renames — no behavioural changes. `cargo build --workspace`,
`cargo test --workspace -- --skip workloads`, `cargo clippy --workspace
--all-targets -- -D warnings`, `cargo run --release -p javm-bench
--example smoke` (12/12 PASS with bit-identical gas).
