+++ rust/javm-bench/benches/bench.rs
