+++ rust/javm-bench/examples/disasm.rs
