+++ rust/javm-bench/examples/smoke.rs
