+++ rust/javm-recompiler-x86/src/lib.rs
@@ -21,7 +21,6 @@ extern crate alloc;
 
 pub mod asm;
 pub mod codegen;
-pub mod rv_codegen;
 
 /// JIT execution context passed to compiled code via R15.
 /// Must be `#[repr(C)]` with exact field ordering matching the
